fix(launch): rename helpers file, register new job in index test, restore mocks
- Rename start-domain-launch-helpers.ts → start-domain-launch-snapshot.ts
  (the file holds configSnapshot parsing helpers; -helpers suffix is
  blocked by the repo's filename policy)
- Update tests/queue/handlers/index.test.ts to expect the new
  start_domain_launch job type (count 13 → 14)
- Add vi.restoreAllMocks() to beforeEach in start-domain-launch.test.ts
  so spy-based test (atomic-claim race-loser) doesn't leak its
  updateMany mock into later tests that exercise the real claim path
