+++ cloud/apps/api/tests/queue/handlers/index.test.ts
@@ -53,7 +53,8 @@ describe('Handler Registration', () => {
       expect(jobTypes).toContain('refresh_domain_analysis_snapshot');
       expect(jobTypes).toContain('refresh_pressure_sensitivity_snapshot');
       expect(jobTypes).toContain('top_up_probes');
-      expect(jobTypes).toHaveLength(13);
+      expect(jobTypes).toContain('start_domain_launch');
+      expect(jobTypes).toHaveLength(14);
     });
   });
 
@@ -104,7 +105,7 @@ describe('Handler Registration', () => {
 
       // Should still work
       const jobTypes = getJobTypes();
-      expect(jobTypes).toHaveLength(13);
+      expect(jobTypes).toHaveLength(14);
     });
   });
