+++ cloud/apps/api/tests/queue/handlers/start-domain-launch.test.ts
@@ -55,6 +55,7 @@ describe('start-domain-launch handler', () => {
   });
 
   beforeEach(() => {
+    vi.restoreAllMocks();
     vi.clearAllMocks();
     getBossMock.mockReturnValue({ getQueues: vi.fn().mockResolvedValue([]) });
     startRunMock.mockImplementation(async (input: StartRunInput) => {
